+++ CHANGELOG.md
@@ -7,6 +7,15 @@ Post v1 a change is further considered breaking if the state file format is inco
 
 ### Upcoming
 
+*Major*
+
+* Added the Necromancer role. ([#211](https://github.com/hjwylde/werewolf/issues/211))
+* Renamed Scapegoat's allowedVoters to chosenVoters. ([#211](https://github.com/hjwylde/werewolf/issues/211))
+
+*Minor*
+
+* Added the Zombie role. ([#211](https://github.com/hjwylde/werewolf/issues/211))
+
 ### v1.3.1.0
 
 *Minor*